backend/requirements.txt
Flask
Flask-Cors
requests
python-dotenv
openai
setuptools